refactor paper detail
