+++ frontend/src/components/Paper/PapersPage.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import styled from "styled-components";
 import { PaperSearchConditionType, PaperType } from "../../types";
-import Papers from "./Papers";
+import Papers from "../Organisms/Papers";
 import SearchBox from "../Search/SearchBox";
 import SearchModal from "../Search/SearchModal";
 import MainButton from "../Atoms/MainButton";
